pyproject.toml
[project]
name = "backend-powermoniter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.8",
    "django-cors-headers>=4.9.0",
    "django-ninja>=1.4.5",
    "paho-mqtt>=2.1.0",
    "prometheus-client>=0.23.1",
    "psycopg[binary]>=3.2.12",
    "pyjwt>=2.10.1",
    "sentry-sdk[django]>=2.43.0",
    "uvicorn[standard]>=0.38.0",
]
